Fix cipher comparison with NID instead of pointers
OpenSSL does in fact _not_ return pointers to static structs; depending
on the state of the library you can get different pointers.

This fixes the internal implementation of `Cipher::is_ccm` and
`Cipher::is_ocb` by doing the comparison with NID instead pointers,
which is a stable identifier of the algorithm, regardless of the state
of the OpenSSL library.