Enhance Debug implementation for Nid
Currently, doing this:

    println!("{:?}", Nid::SECP521R1);

outputs:

    Nid(716)

Which is not as helpful as it could be.

This commit switches Nid over to a manual implementation of Debug which
prints the Nid's integer value, and, additionally, the Nid's short name
if it has one.  This changes the output of the command above to:

    Nid { nid: 716, short_name: "secp521r1" }

And the output of this:

    println!("{:?}", Nid::from_raw(123456789));

becomes:

    Nid { nid: 123456789 }